# apidoc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/book-catalog": {
      "post": {
        "description": "Crea un nuevo libro en el catálogo del sistema. Solo accesible para administradores.",
        "operationId": "BookCatalogController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookCatalogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Libro creado exitosamente en el catálogo",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BookCatalogResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos o errores de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token JWT inválido o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado - Se requieren permisos de administrador",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "El código ISBN ya existe en el catálogo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear nuevo libro en el catálogo - Acceso: ADMIN",
        "tags": [
          "Book Catalog"
        ]
      },
      "get": {
        "description": "Obtiene una lista paginada de todos los libros en el catálogo con filtros opcionales. Endpoint público.",
        "operationId": "BookCatalogController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Término de búsqueda para filtrar por título del libro",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genreId",
            "required": false,
            "in": "query",
            "description": "ID del género para filtrar libros por categoría",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "publisherId",
            "required": false,
            "in": "query",
            "description": "ID de la editorial para filtrar libros por publicador",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAvailable",
            "required": false,
            "in": "query",
            "description": "Filtrar por disponibilidad del libro",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Precio mínimo para filtrar libros",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Precio máximo para filtrar libros",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo de libros obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookCatalogResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "page": {
                                  "type": "number"
                                },
                                "limit": {
                                  "type": "number"
                                },
                                "totalPages": {
                                  "type": "number"
                                },
                                "hasNext": {
                                  "type": "boolean"
                                },
                                "hasPrev": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Obtener catálogo de libros - Acceso: ADMIN, USER",
        "tags": [
          "Book Catalog"
        ]
      }
    },
    "/api/book-catalog/search": {
      "get": {
        "description": "Busca libros en el catálogo por un término de búsqueda en el título, código ISBN o resumen. Endpoint público.",
        "operationId": "BookCatalogController_search",
        "parameters": [
          {
            "name": "term",
            "required": true,
            "in": "query",
            "description": "Término de búsqueda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de la búsqueda de libros",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookCatalogResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "page": {
                                  "type": "number"
                                },
                                "limit": {
                                  "type": "number"
                                },
                                "totalPages": {
                                  "type": "number"
                                },
                                "hasNext": {
                                  "type": "boolean"
                                },
                                "hasPrev": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token JWT inválido o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar libros en el catálogo - Acceso: ADMIN, USER",
        "tags": [
          "Book Catalog"
        ]
      }
    },
    "/api/book-catalog/filter": {
      "post": {
        "description": "Filtra libros del catálogo utilizando criterios avanzados como género, precio, disponibilidad, etc.",
        "operationId": "BookCatalogController_filter",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Criterios de filtrado para los libros",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookFiltersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Libros filtrados obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookCatalogResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "page": {
                                  "type": "number"
                                },
                                "limit": {
                                  "type": "number"
                                },
                                "totalPages": {
                                  "type": "number"
                                },
                                "hasNext": {
                                  "type": "boolean"
                                },
                                "hasPrev": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token JWT inválido o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Filtrar libros con criterios avanzados - Acceso: ADMIN, USER",
        "tags": [
          "Book Catalog"
        ]
      }
    },
    "/api/book-catalog/available": {
      "get": {
        "description": "Obtiene una lista paginada de libros disponibles para compra en el catálogo.",
        "operationId": "BookCatalogController_findAvailable",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Libros disponibles obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookCatalogResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "page": {
                                  "type": "number"
                                },
                                "limit": {
                                  "type": "number"
                                },
                                "totalPages": {
                                  "type": "number"
                                },
                                "hasNext": {
                                  "type": "boolean"
                                },
                                "hasPrev": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token JWT inválido o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener libros disponibles - Acceso: ADMIN, USER",
        "tags": [
          "Book Catalog"
        ]
      }
    },
    "/api/book-catalog/by-genre/{genreId}": {
      "get": {
        "description": "Obtiene una lista paginada de libros que pertenecen a un género específico.",
        "operationId": "BookCatalogController_findByGenre",
        "parameters": [
          {
            "name": "genreId",
            "required": true,
            "in": "path",
            "description": "ID del género para filtrar los libros",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Libros del género especificado obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookCatalogResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "page": {
                                  "type": "number"
                                },
                                "limit": {
                                  "type": "number"
                                },
                                "totalPages": {
                                  "type": "number"
                                },
                                "hasNext": {
                                  "type": "boolean"
                                },
                                "hasPrev": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token JWT inválido o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener libros por género - Acceso: ADMIN, USER",
        "tags": [
          "Book Catalog"
        ]
      }
    },
    "/api/book-catalog/by-publisher/{publisherId}": {
      "get": {
        "description": "Obtiene una lista paginada de libros que pertenecen a una editorial específica.",
        "operationId": "BookCatalogController_findByPublisher",
        "parameters": [
          {
            "name": "publisherId",
            "required": true,
            "in": "path",
            "description": "ID de la editorial para filtrar los libros",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Libros de la editorial especificada obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookCatalogResponseDto"
                              }
                            },
                            "meta": {
                              "type": "object",
                              "properties": {
                                "total": {
                                  "type": "number"
                                },
                                "page": {
                                  "type": "number"
                                },
                                "limit": {
                                  "type": "number"
                                },
                                "totalPages": {
                                  "type": "number"
                                },
                                "hasNext": {
                                  "type": "boolean"
                                },
                                "hasPrev": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token JWT inválido o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener libros por editorial - Acceso: ADMIN, USER",
        "tags": [
          "Book Catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Store API Documentation",
    "description": "Store API is a simple API that allows you to manage books.",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateBookCatalogDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the book",
            "example": "The Shining",
            "minLength": 1,
            "maxLength": 255
          },
          "isbnCode": {
            "type": "string",
            "description": "Unique ISBN code for the book",
            "example": "9780307743657",
            "minLength": 10,
            "maxLength": 13
          },
          "price": {
            "type": "number",
            "description": "Price of the book",
            "example": 19.99,
            "minimum": 0,
            "format": "decimal"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Indicates if the book is available for purchase",
            "example": true,
            "default": true
          },
          "stockQuantity": {
            "type": "number",
            "description": "Available stock quantity of the book",
            "example": 25,
            "minimum": 0,
            "default": 0
          },
          "coverImageUrl": {
            "type": "string",
            "description": "URL of the book cover image",
            "example": "https://example.com/covers/the-shining.jpg",
            "format": "url"
          },
          "publicationDate": {
            "format": "date",
            "type": "string",
            "description": "Date when the book was published",
            "example": "1947-09-21"
          },
          "pageCount": {
            "type": "number",
            "description": "Number of pages in the book",
            "example": 447,
            "minimum": 1
          },
          "summary": {
            "type": "string",
            "description": "Summary or description of the book",
            "example": "A psychological horror novel that tells the story of Jack Torrance, an aspiring writer and recovering alcoholic who accepts a position as the off-season caretaker of the isolated historic Overlook Hotel."
          },
          "genreId": {
            "type": "string",
            "description": "Foreign key reference to book genre",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "format": "uuid"
          },
          "publisherId": {
            "type": "string",
            "description": "Foreign key reference to publishing house",
            "example": "550e8400-e29b-41d4-a716-446655440001",
            "format": "uuid"
          }
        },
        "required": [
          "title",
          "isbnCode",
          "price",
          "genreId",
          "publisherId"
        ]
      },
...
    }
  }
}